requirements.txt
smartapi-python
requests
schedule
python-dotenv
python-telegram-bot
